2 items changed
  id: recFriRxcb5aA4Crq
    Total reports: "5" => "6"
    Yeses: "1" => "2"

  id: recgC68ghhJW0EDz4
    Notes: "[Feb 23] Vaccinating Phase Phase 1A health care sector, long-term care patients, 75 and older and (as of Feb 9th) 65 and older.

The County of Ventura will also begin vaccinating teachers and child care providers, first responders and food and agricultural workers who are eligible under the State's Phase 1B as supply allows March 1.

The supply of COVID-19 Vaccines is limited. Appointments open each week.

**To make an appointment:**

To make an appointment, visit the COVID-19 Vaccination Registration Portal at [www.venturacountyrecovers.org/portal/](https://www.venturacountyrecovers.org/vaccine-information/portal/).

Community members who are eligible to be vaccinated but who have limited computer/internet access may call the registration hotline on Monday's to make an appointment. The phone number is: 805-477-7161.

The Area Agency on Aging provides transportation services for older adults that are in need of transportation to a vaccine appointment or for COVID-19 testing. This includes people that need lift assist as well as gurney transportation. The driver will take the individual, wait for them, help them at the site and take them home. To schedule a ride call 805-477-7300

Register for vaccine updates with the county at <https://www.venturacountyrecovers.org/vaccine-information/subscribe-vaccine/>.

Individuals who are eligible to be vaccinated are encouraged to contact their personal healthcare provider for information for vaccine information and to check for availability.

" => "[Feb 28] Currently vaccinating the following groups:

- Individuals aged 65 and older
- Healthcare workers
- Long-term care residents
- Essential workers in the following sectors: Food and Agriculture, Education and Childcare, Emergency Services
- Tri-Counties Regional Centers. Family member caregivers with direct risk of COVID-19 exposure who care for regional center consumers at high risk of COVID-19 complications

All food and agricultural workers are now able to make appointments for the COVID-19 vaccine. More appointments will open Monday morning (March 1st) at 7 am online and 8 am on the phone line for those who do not have access to the internet 805-477-7161.

The supply of COVID-19 Vaccines is limited. Appointments open each week.

**To make an appointment:**

To make an appointment, visit the COVID-19 Vaccination Registration Portal at [www.venturacountyrecovers.org/portal/](https://www.venturacountyrecovers.org/vaccine-information/portal/).

Community members who are eligible to be vaccinated but who have limited computer/internet access may call the registration hotline on Monday's to make an appointment. The phone number is: 805-477-7161.

The Area Agency on Aging provides transportation services for older adults that are in need of transportation to a vaccine appointment or for COVID-19 testing. This includes people that need lift assist as well as gurney transportation. The driver will take the individual, wait for them, help them at the site and take them home. To schedule a ride call 805-477-7300

Register for vaccine updates with the county at <https://www.venturacountyrecovers.org/vaccine-information/subscribe-vaccine/>.

Individuals who are eligible to be vaccinated are encouraged to contact their personal healthcare provider for information for vaccine information and to check for availability.

"