1 item changed
  id: recFriRxcb5aA4Crq
    Notes: "[March 5] Currently vaccinating individuals  in Phase 1A- all Tiers and Phase 1B, which includes those age 65 and older, and those working in
childcare and education. emergency services, food and agriculture..

If you are interested in receiving the vaccine, you can fill out
[this form](https://forms.office.com/Pages/ResponsePage.aspx?id=fCdnJLfxW0qLdexgCSM17lXbsjwkAmVKlpPwkl1gWP5UODdLVTY3RlhLMk5PRk5VVU1VRUxaMEhFQi4u) to be placed on the County contact list. The team is working through the COVID-19 Vaccine Contact List to place individuals into their appropriate Phases and Tiers, setting priority vaccinations based on risks such as age and underlying conditions.
Appointments are required for all vaccines.

**When the vaccine becomes available for your age group, you will receive a call to schedule your appointment. **

If you do not have access to the internet, please call (209) 259-1332 for assistance. 

If you want an update on where the county is currently in the age groups, please call **(209) 966-1133** and listen to the whole recording for the most up-to-date details. 

For more information you can also:
Visit the county's Facebook page at <https://www.facebook.com/MariposaCountyPublicHealth/?ref=bookmarks>.    or
Visit the county's Instagram page at <https://www.instagram.com/mariposacounty_hhsa/>.
" => "[March 5] Currently vaccinating individuals  in Phase 1A- all Tiers and Phase 1B, which includes those age 65 and older, and those working in
childcare and education. emergency services, food and agriculture..

If you are interested in receiving the vaccine, you can fill out
[this form](https://forms.office.com/Pages/ResponsePage.aspx?id=fCdnJLfxW0qLdexgCSM17lXbsjwkAmVKlpPwkl1gWP5UODdLVTY3RlhLMk5PRk5VVU1VRUxaMEhFQi4u) to be placed on the County contact list. The team is working through the COVID-19 Vaccine Contact List to place individuals into their appropriate Phases and Tiers, setting priority vaccinations based on risks such as age and underlying conditions.
Appointments are required for all vaccines.
**When the vaccine becomes available for your age group, you will receive a call to schedule your appointment. **

If you do not have access to the internet, please call** (209) 259-1332** for assistance. 

If you want an update on where the county is currently in the age groups, please call **(209) 966-1133** and listen to the whole recording for the most up-to-date details. 

For more information you can also:
Visit the county's Facebook page at <https://www.facebook.com/MariposaCountyPublicHealth/?ref=bookmarks>.    or
Visit the county's Instagram page at <https://www.instagram.com/mariposacounty_hhsa/>.
"