1 item changed
  id: recipKOzQdBxv0mAf
    Notes: "[March 5] Currently vaccinating Individuals age  65 and older, Phase 1A- all Tiers, and Phase 1B-Tier 1  For more detailed information about who qualifies, please go here: [Vaccine Plan & Distribution in Tulare County - COVID 19 (ca.gov)](https://covid19.tularecounty.ca.gov/covid-19-vaccine/vaccine-plan-distribution-in-tulare-county/)

To view locations where the vaccine is available and register for an appointment if in an eligible group, go here:
[COVID-19 Vaccine - COVID 19 (ca.gov)](https://covid19.tularecounty.ca.gov/covid-19-vaccine/)

For those without internet access can call 1-833-422-4255 to make an appointment.

To be notified when you are eligible to receive your vaccine, and to register, sign up here:
[My Turn - California COVID-19 Vaccine Scheduling & Notifications](https://myturn.ca.gov/)
or call 833-422-4255 to make an appointment.

If you are interested in volunteering to help Tulare County with the vaccine rollout, you can sign up here: [Calling All Volunteers! - COVID 19](https://covid19.tularecounty.ca.gov/covid-19-vaccine/calling-all-volunteers/)
" => "(**Updated:** Mar 6)  **Eligibility:** 65+, or in Emergency services, Education & Childcare, and Food and Agriculture workers.

**County restrictions:** Must be a Tulare county resident or work in Tulare county.

**Appointment Information:**
To view locations where the vaccine is available and register for an appointment if in an eligible group, go here:
[COVID-19 Vaccine - COVID 19 (ca.gov)](https://covid19.tularecounty.ca.gov/covid-19-vaccine/)

To be notified when you are eligible to receive your vaccine, and to register, sign up here:
[My Turn - California COVID-19 Vaccine Scheduling & Notifications](https://myturn.ca.gov/)

In addition, there is a complete list of local healthcare providers currently offering COVID vaccinations on the County webpage here:
<https://covid19.tularecounty.ca.gov/covid-19-vaccine/make-a-vaccination-appointment/covid-vaccination-sites-in-tulare-county/>

**Phone numbers:** Individuals without access to the internet can call the MyTurn Hotline at 1-833-422-4255 to make an appointment.

**Additional Info:**
"