4 items changed
  id: recb6ibrCGCw71RbX
    Notes: "(Updated: Mar 10)  **Eligibility:** 65+, or 16+ who work in Emergency services, Education & Childcare, and Food and Agriculture workers

Note: 16 &17 year olds who work in these sectors are only eligible to receive the Pfizer vaccine.

**County restrictions:** Must be a Stanislaus county resident or work in Stanislaus county

**Appointment Information:**
The county has details on clinics (mostly first-come, first-served, ie no appointments needed) on their website at <http://schsa.org/coronavirus/vaccine/>

**Phone numbers:** Public Information line at (209) 558-7535

**Additional Info:** County site updated every Friday or as vaccine availability changes.

According to the county, vaccination screening forms will be provided at the clinics. If you wish to fill out your screening prior to ariving at the clinic, the form is available for you to print at <http://schsa.org/coronavirus/vaccine/pdf/covid-vaccine-screening.pdf>
" => "(Updated: Mar 11)  **Eligibility:** 65+, or 16+ who work in Emergency services, Education & Childcare, and Food and Agriculture workers

Note: 16 &17 year olds who work in these sectors are only eligible to receive the Pfizer vaccine.

**County restrictions:** Must be a Stanislaus county resident or work in Stanislaus county

**Appointment Information:**
The county has details on clinics (mostly first-come, first-served, ie no appointments needed) on their website at <http://schsa.org/coronavirus/vaccine/>

**Phone numbers:** Public Information line at (209) 558-7535

**Additional Info:** County site updated every Friday or as vaccine availability changes.

According to the county, vaccination screening forms will be provided at the clinics. If you wish to fill out your screening prior to arriving at the clinic, the form is available for you to print at <http://schsa.org/coronavirus/vaccine/pdf/covid-vaccine-screening.pdf>
"

  id: recipKOzQdBxv0mAf
    Notes: "(Updated: Mar 10)  **Eligibility:** 65+, or in Emergency services, Education & Childcare, and Food and Agriculture workers.

**County restrictions:** Must be a Tulare county resident or work in Tulare county.

**Appointment Information:**
To view locations where the vaccine is available and register for an appointment if in an eligible group, go here:
[COVID-19 Vaccine - COVID 19 (ca.gov)](https://covid19.tularecounty.ca.gov/covid-19-vaccine/)

To be notified when you are eligible to receive your vaccine, and to register, sign up here:
[My Turn - California COVID-19 Vaccine Scheduling & Notifications](https://myturn.ca.gov/)

In addition, there is a complete list of local healthcare providers currently offering COVID vaccinations on the County webpage here:
<https://covid19.tularecounty.ca.gov/covid-19-vaccine/make-a-vaccination-appointment/covid-vaccination-sites-in-tulare-county/>

**Phone numbers:** Individuals without access to the internet can call the MyTurn Hotline at 1-833-422-4255 to make an appointment.

**Additional Info: **Tulare County residents and workers who are currently eligible to receive the COVID-19 vaccine and have been unable to secure an appointment elsewhere can now get vaccinated at clinics located at the Lindsay Wellness Center and Earlimart Memorial Building. Visit My Turn at: <https://myturn.ca.gov/> to sign up, or call 1-833-422-4255 if in need of assistance over the phone.
" => "(Updated: Mar 11)  **Eligibility:** 65+, or in Emergency services, Education & Childcare, and Food and Agriculture workers.

**County restrictions:** Must be a Tulare county resident or work in Tulare county.

**Appointment Information:**
To view locations where the vaccine is available and register for an appointment if in an eligible group, go here:
[COVID-19 Vaccine - COVID 19 (ca.gov)](https://covid19.tularecounty.ca.gov/covid-19-vaccine/)

To be notified when you are eligible to receive your vaccine, and to register, sign up here:
[My Turn - California COVID-19 Vaccine Scheduling & Notifications](https://myturn.ca.gov/)

In addition, there is a complete list of local healthcare providers currently offering COVID vaccinations on the County webpage here:
<https://covid19.tularecounty.ca.gov/covid-19-vaccine/make-a-vaccination-appointment/covid-vaccination-sites-in-tulare-county/>

**Phone numbers:** Individuals without access to the internet can call the MyTurn Hotline at 1-833-422-4255 to make an appointment.

**Additional Info: **Tulare County residents and workers who are currently eligible to receive the COVID-19 vaccine and have been unable to secure an appointment elsewhere can now get vaccinated at clinics located at the Lindsay Wellness Center and Earlimart Memorial Building. Visit My Turn at: <https://myturn.ca.gov/> to sign up, or call 1-833-422-4255 if in need of assistance over the phone.
"

  id: recl8GEIaG1m1qokG
    Notes: "(Updated: Mar 10)  **Eligibility**: 65+, or in Education, Emergency services, Childcare, and Food and Agriculture workers

**County restrictions**: Must be a Merced county resident or work in Merced county

**Appointment Information**: For locations and appointment information please click this link: <https://vaccinatemercedcounty.com/registration/>

**Phone numbers**:

**Additional Info**:
" => "(Updated: Mar 11)  **Eligibility**: 65+, or in Education, Emergency services, Childcare, and Food and Agriculture workers

**County restrictions**: Must be a Merced county resident or work in Merced county

**Appointment Information**: For locations and appointment information please click this link: <https://vaccinatemercedcounty.com/registration/>

**Phone numbers**:

**Additional Info**:
"

  id: recyKBuA9lxrJc339
    Notes: "(Updated: Mar 10)  **Eligibility:** 65+, essential workers, and those 18-
64 with underlying medical conditions.

**County restrictions:** Must be an Alpine county resident (whether full-time resident OR part-time resident & property owner) or work in Alpine county.

**Appointment Information:**
As of Feb 24th, County is requesting anyone 18+ (that is eligible as per the "county restrictions" section above) to call the Warm Line below and register their interest. A list will be complied, and names will be prioritized based on age, underlying medical conditions,
occupation, and residency.

- Woodfords and Bear Valley – call the Warm Line at 530-694-1011
- Washoe Tribal Health Center – call 775-265-8622

**Phone numbers:** Individuals without access to the internet can call the Warm Line at 530-694-1011.

**Additional Info:** <https://alpinecountyca.gov/DocumentCenter/View/4272/ACPHB-vaccine-eligibility-update-22421?bidId=>

" => "(Updated: Mar 11)  **Eligibility:** 65+, essential workers, and those 18-
64 with underlying medical conditions.

**County restrictions:** Must be an Alpine county resident (whether full-time resident OR part-time resident & property owner) or work in Alpine county.

**Appointment Information:**
As of Feb 24th, County is requesting anyone 18+ (that is eligible as per the "county restrictions" section above) to call the Warm Line below and register their interest. A list will be complied, and names will be prioritized based on age, underlying medical conditions,
occupation, and residency.

- Woodfords and Bear Valley – call the Warm Line at 530-694-1011
- Washoe Tribal Health Center – call 775-265-8622

**Phone numbers:** Individuals without access to the internet can call the Warm Line at 530-694-1011.

**Additional Info:** <https://alpinecountyca.gov/DocumentCenter/View/4272/ACPHB-vaccine-eligibility-update-22421?bidId=>
"