1 item changed
  id: recyKBuA9lxrJc339
    Notes: "(Updated: Mar 19)  **Eligibility:** 18+

**County restrictions: Must be an Alpine county resident (whether full-time resident OR part-time resident & property owner) or essential worker in Alpine county.**

**Appointment Information:**
As of Feb 24th, County is requesting anyone 18+ (that is eligible as per the "county restrictions" section above) to call the Warm Line below and register their interest. A list will be complied, and names will be prioritized based on age, underlying medical conditions,
occupation, and residency.

- Woodfords and Bear Valley – call the Warm Line at 530-694-1011
- Washoe Tribal Health Center – call 775-265-8622

**Phone numbers:** Individuals without access to the internet can call the Warm Line at 530-694-1011.

**Additional Info:** <https://alpinecountyca.gov/DocumentCenter/View/4272/ACPHB-vaccine-eligibility-update-22421?bidId=>
" => "(Updated: Mar 28)  **Eligibility:** 18+

**County restrictions: Must be an Alpine county resident (whether full-time resident OR part-time resident & property owner) or essential worker in Alpine county.**

**Appointment Information:**
As of Feb 24th, County is requesting anyone 18+ (that is eligible as per the "county restrictions" section above) to call the Warm Line below and register their interest. A list will be complied, and names will be prioritized based on age, underlying medical conditions,
occupation, and residency.

- Woodfords and Bear Valley – call the Warm Line at 530-694-1011
- Washoe Tribal Health Center – call 775-265-8622

**Phone numbers:** Individuals without access to the internet can call the Warm Line at 530-694-1011.

**Additional Info:** <https://alpinecountyca.gov/DocumentCenter/View/4272/ACPHB-vaccine-eligibility-update-22421?bidId=>
"