2 items changed
  id: rec2V4GnfzE4qF2a0
    Yeses: "122" => "123"

  id: reclZ8DWOEuoluStG
    Notes: "(Updated: Apr 5) **Eligibility**: 50+, or education, emergency services and food and agriculture workers. Persons age 16-64 with serious underlying medical conditions or disabilities. Also eligible: public transit workers, utility workers, foster parents, social workers, disaster service workers, those who serve the homeless, airport workers, and janitors.  [Details on eligibility here.](https://www.cdph.ca.gov/Programs/CID/DCDC/Pages/COVID-19/vaccine-high-risk-factsheet.aspx)

Also: Effective immediately, those who accompany an eligible family member to an appointment are eligible, too, even if they don’t fall into an eligible category – but they must have an appointment at the same location and at approximately the same time as their eligible family member. Those accompanying someone 65 or over, a person with a serious underlying health condition or disability, a food worker, or anyone else who is eligible should sign up for an appointment as a member of that eligible cohort, and accompany that eligible person to the appointment.

**County restrictions**: Must live or work in San Bernardino County

**Appointment Information**: <https://sbcovid19.com/vaccine/locations/>

**Phone number**: For vaccine questions and help, call the county's COVID-19 hotline at (909) 387-3911 from 9 a.m. to 8 p.m. Monday through Friday.

**Additional information**: While no doctor's notes are required for those who claim eligibility because of a medical condition, you will need to show a signed self-attestation form at your appointment. [Here is the form](https://sbcovid19.com/wp-content/uploads/sites/41/2021/03/Self-Attestation.pdf).
Omnitrans offers [free vaccination shuttles](https://omnitrans.org/blog/omnitrans-offers-free-vaccination-shuttle-in-ontario-and-san-bernardino/) in Ontario and San Bernardino.
###

" => "(Updated: Apr 6) **Eligibility**: All people 16 or older.
Also: Effective immediately, those who accompany an eligible family member to an appointment are eligible, too, even if they don’t fall into an eligible category – but they must have an appointment at the same location and at approximately the same time as their eligible family member. Those accompanying someone 65 or over, a person with a serious underlying health condition or disability, a food worker, or anyone else who is eligible should sign up for an appointment as a member of that eligible cohort, and accompany that eligible person to the appointment.

**County restrictions**: Must live or work in San Bernardino County

**Appointment Information**: <https://sbcovid19.com/vaccine/locations/>

**Phone number**: For vaccine questions and help, call the county's COVID-19 hotline at (909) 387-3911 from 9 a.m. to 8 p.m. Monday through Friday.

Omnitrans offers [free vaccination shuttles](https://omnitrans.org/blog/omnitrans-offers-free-vaccination-shuttle-in-ontario-and-san-bernardino/) in Ontario and San Bernardino.
###

"