4 items changed
  id: rec7vkrMOgPeUg6Z2
    Yeses: "52" => "51"

  id: recarUDlLAO0MtvA7
    Yeses: "285" => "286"

  id: reclZ8DWOEuoluStG
    Notes: "(Updated: Apr 6) **Eligibility**: All people 16 or older.
Also: Effective immediately, those who accompany an eligible family member to an appointment are eligible, too, even if they don’t fall into an eligible category – but they must have an appointment at the same location and at approximately the same time as their eligible family member. Those accompanying someone 65 or over, a person with a serious underlying health condition or disability, a food worker, or anyone else who is eligible should sign up for an appointment as a member of that eligible cohort, and accompany that eligible person to the appointment.

**County restrictions**: Must live or work in San Bernardino County

**Appointment Information**: <https://sbcovid19.com/vaccine/locations/>

**Phone number**: For vaccine questions and help, call the county's COVID-19 hotline at (909) 387-3911 from 9 a.m. to 8 p.m. Monday through Friday.

Omnitrans offers [free vaccination shuttles](https://omnitrans.org/blog/omnitrans-offers-free-vaccination-shuttle-in-ontario-and-san-bernardino/) in Ontario and San Bernardino.
###

" => "(Updated: Apr 6) **Eligibility**: 16+
Also: Effective immediately, those who accompany an eligible family member to an appointment are eligible, too, even if they don’t fall into an eligible category – but they must have an appointment at the same location and at approximately the same time as their eligible family member. Those accompanying someone 65 or over, a person with a serious underlying health condition or disability, a food worker, or anyone else who is eligible should sign up for an appointment as a member of that eligible cohort, and accompany that eligible person to the appointment.

**County restrictions**: Must live or work in San Bernardino County

**Appointment Information**: <https://sbcovid19.com/vaccine/locations/>

**Phone number**: For vaccine questions and help, call the county's COVID-19 hotline at (909) 387-3911 from 9 a.m. to 8 p.m. Monday through Friday.

Omnitrans offers [free vaccination shuttles](https://omnitrans.org/blog/omnitrans-offers-free-vaccination-shuttle-in-ontario-and-san-bernardino/) in Ontario and San Bernardino.
###

"

  id: recmEBjsHW6t9GDHg
    Notes: "(Updated: Apr 6) **Eligibility**:** **18+

**County restrictions**: Must live or work full-time in Lassen County

**Appointment Information**: Submit an appointment request form, available at <https://lassencares.org/>

**Phone number**: You can also call (530) 249-8628 to submit your appointment request information.

" => "(Updated: Apr 6) **Eligibility**: 18+
**County restrictions**: Must live or work full-time in Lassen County
**Appointment Information**: Submit an appointment request form, available at <https://lassencares.org/>
**Phone number**: You can also call (530) 249-8628 to submit your appointment request information.

"